+++ Sources/SwiftKit/_Deprecated/ApiServiceError.swift
@@ -1,0 +1,2 @@
+@available(*, deprecated, renamed: "ApiError")
+public typealias ApiServiceError = ApiError