+++ Release Notes.md
@@ -3,9 +3,27 @@
 I will bump revision by revision, until SwiftKit has all functionality that it should have from iExtra. I will then bump it to `1.0.0`.
 
 
+## 0.5.0
+
+
+### New features
+
+* `ApiError` is a new enum that replaces the old `ApiServiceError`.
+
+### Deprecations
+
+* `ApiServiceError` has been deprecated in favor of `ApiError`.
+
+### Breaking changes
+
+* `ApiError` is a lot easier than before, with many cases gone and more info in the remaining ones.
+
+
+
 ## 0.4.4
 
 This version adds a new `HttpMethod` enum that can be used with the network components.
+
 
 
 ## 0.4.3
@@ -13,9 +31,11 @@ This version adds a new `HttpMethod` enum that can be used with the network comp
 This version adds new `ApiRoute` request functions and adjusts the url of form data requests.
 
 
+
 ## 0.4.2
 
 This version adds missing initializers to map services.
+
 
 
 ## 0.4.1
@@ -23,9 +43,11 @@ This version adds missing initializers to map services.
 This version updates dependencies, adjusts project setup, tweak icons and display names etc.
 
 
+
 ## 0.4.0
 
 This version adds a new `String+Dictation` extension to cleanup dictation objects and spaces from a string.
+
 
 
 ## 0.3.3
@@ -35,9 +57,11 @@ This version adds a new `ExternalMapService` protocol as well as an Apple and a 
 This version also adds a new `Network` namespace, with api-specific protocols and errors for communicating with external REST apis.
 
 
+
 ## 0.3.2
 
 This version adds a `UserDefaults+Codable` extension for persisting codable types in `UserDefaults`.
+
 
 
 ## 0.3.1
@@ -45,11 +69,13 @@ This version adds a `UserDefaults+Codable` extension for persisting codable type
 This version makes the standard cvs parser use paths instead of urls when parsing files.
 
 
+
 ## 0.3.0
 
 This version adds improved support for watchOS and tvOS.
 
 The bump version process has been improved to also add linting and a unit test confirmation.
+
 
 
 ## 0.2.0
@@ -62,6 +88,7 @@ This version adds:
 This version also adds macOS support.
 
 
+
 ## 0.1.0
 
 This version adds:
@@ -71,9 +98,11 @@ This version adds:
 * new `Numeric` conversion [extensions][Extensions].
 
 
+
 ## 0.0.6
 
 This version adds a bunch of [extensions][Extensions] and common utils and updates external test dependencies to the latest versions.
+
 
 
 ## 0.0.5
@@ -81,9 +110,11 @@ This version adds a bunch of [extensions][Extensions] and common utils and updat
 This version adds a bunch of [extensions][Extensions] and common types.
 
 
+
 ## 0.0.4
 
 This version adds [device][Device] and [keychain][Keychain] utils.
+
 
 
 ## 0.0.3
@@ -91,9 +122,11 @@ This version adds [device][Device] and [keychain][Keychain] utils.
 This version adds a bunch of convenient [extensions][Extensions]. 
 
 
+
 ## 0.0.2
 
 This version adjusts [authentication][Authentication] service signatures, as well as [coding][Coding] and [IoC][IoC] functionality  
+
 
 
 ## 0.0.1